requirements.txt
# Anti-Deepfake Vision FGSM Imputer
# Python dependencies

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Data augmentation
albumentations>=1.3.0

# Deep learning utilities
tqdm>=4.65.0
pyyaml>=6.0

# Visualization and monitoring
tensorboard>=2.13.0
matplotlib>=3.7.0

# Web interface
gradio>=4.0.0

# Export formats
onnx>=1.14.0
onnxruntime>=1.15.0

# Development tools (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Google Colab support
ipython>=8.12.0
ipykernel>=6.23.0
